+++ examples/write_html.rs
@@ -1,16 +1,15 @@
 extern crate ytml;
 
 use std::collections::HashMap;
-use ytml::{file_handling::file_output::write_html_to_file, ast::Tag};
+use ytml::{file_handling::file_output::write_html_to_file, tokens::Tag};
 
 fn main() {
-    let document = vec![
-        Tag{
-            attributes: HashMap::new(),
-            inner: Vec::new(),
-            name: String::from("html"),
-        }
-    ];
+    let document = vec![Tag {
+        attributes: HashMap::new(),
+        inner: Vec::new(),
+        name: String::from("html"),
+    }];
     let file_path = "./out.html";
     write_html_to_file(file_path, document, 2);
 }
+
