+++ src/lib.rs
@@ -1,4 +1,4 @@
-pub mod ast;
+pub mod tokens;
 pub mod file_handling;
 pub mod html;
 pub mod ytml;
